Allow to set jsonMemberNames on enum, affecting all cases
This change allows to set `@jsonMemberNames` annotation on
enum (sealed trait) to affect all cases, so it is no longer necessary
to add the annotation to each of the cases individually. Each case
can, however, override it.

Basic tests were added verifying that the annotation has effect and
cases can override enum's annotation.